+++ src/features/calendar/contexts/calendar-context/context.ts
@@ -4,7 +4,7 @@ import type { CellClickInfo } from '@/features/calendar/types'
 import type { RecurrenceEditOptions } from '@/features/recurrence/types'
 import type dayjs from '@/lib/configs/dayjs-config'
 import type { TranslationKey } from '@/lib/translations/types'
-import type { CalendarView } from '@/types'
+import type { CalendarView, TimeFormat } from '@/types'
 import { createContext, useContext } from 'react'
 
 export interface CalendarContextType {
@@ -57,7 +57,7 @@ export interface CalendarContextType {
   renderEventForm?: (props: EventFormProps) => React.ReactNode
   // Translation function
   t: (key: TranslationKey) => string
-  is24Hour: boolean
+  timeFormat: TimeFormat
 }
 
 export const CalendarContext: React.Context<CalendarContextType | undefined> =