+++ src/features/resource-calendar/components/ilamy-resource-calendar/ilamy-resource-calendar.tsx
@@ -38,7 +38,7 @@ export const IlamyResourceCalendar: React.FC<IlamyResourceCalendarProps> = ({
   renderResource,
   renderEventForm,
   businessHours,
-  is24Hour = false,
+  timeFormat = '12-hour',
 }) => {
   return (
     <ResourceCalendarProvider
@@ -72,7 +72,7 @@ export const IlamyResourceCalendar: React.FC<IlamyResourceCalendarProps> = ({
       translator={translator}
       renderEventForm={renderEventForm}
       businessHours={businessHours}
-      is24Hour={is24Hour}
+      timeFormat={timeFormat}
     >
       <ResourceCalendarBody />
     </ResourceCalendarProvider>