+++ src/index.ts
@@ -24,7 +24,7 @@ export type { IlamyResourceCalendarProps } from './features/resource-calendar/co
 export type { WeekDays } from './components/types'
 export type { UseIlamyCalendarContextReturn } from './features/calendar/contexts/calendar-context/context'
 export type { UseIlamyResourceCalendarContextReturn } from './features/resource-calendar/contexts/resource-calendar-context/context'
-export type { CalendarView } from './types'
+export type { CalendarView, TimeFormat } from './types'
 
 // Resource calendar types
 export type { Resource } from './features/resource-calendar/types'